References and Examples on briannesbitt/carbon and morilog/jalali
Added Reference and Examples on how to use briannesbitt/carbon with morilog/jalali and where to find useful methods. It took me a lot to figure these out and finally found it in issues so If you please let's add it for people to be able to find it faster.